All the best, Jacopo!
It was a pleasure working with you all those months, and as you
were my first (albeit unofficial) supervisor, I am especially
grateful for you welcoming me to the team. As sad as it might be
to see you go, I sincerely hope that wherever you'll end up,
you'll be appreciated as you're definitely one of the smartest
humans I've met (:

So I wish you all the best, and also concur what Dinika said:
don't be a stranger... and we need to see that youtube channel!

Signed-off-by: Szymon Łopaciuk <szymon@lopaciuk.eu>