Hey man,
It was great working with you, even though it was a very short period.
We never did the hackathon for workflows :P but don't worry we'll replace
them soon!

I wish you all the very best for the future and hope you'll
invite me for a tour at Google.

Best,
Harris