jacquerie: put module in its own package
The jacquerie module has been part of inspire-next for a long time and
has been a foundation of many of its core functionalities.  Its builtin
linter had a tendency of being very strict and printing angry messages.
However, it also provided very valuable and entertaining information and
will be missed.

As an independent package, the module will hopefully be maintained by a
new team so that it can be deployed in a proper environment, either
on premise or in the cloud.

In the future, the module will be rewritten in C++ on a mechanical
keyboard using algorithms with optimal complexity.

Signed-off-by: Micha Moskovic <michamos@gmail.com>